Send an IAC to the i960 instead of some goofy protocol
We can just encode IAC packets to it.

Also, use timer 1 instead of timer 2 so I have more fine grained control
over the timer frequency

+++ Chipset/include/TimerDevice.h
@@ -43,9 +43,9 @@ public:
     bool begin() noexcept ;
     bool isAvailable() const noexcept { return true; }
     [[nodiscard]] uint8_t getSystemTimerPrescalar() const noexcept;
-    [[nodiscard]] uint8_t getSystemTimerComparisonValue() const noexcept;
+    [[nodiscard]] uint16_t getSystemTimerComparisonValue() const noexcept;
     void setSystemTimerPrescalar(uint8_t value) noexcept;
-    void setSystemTimerComparisonValue(uint8_t value) noexcept;
+    void setSystemTimerComparisonValue(uint16_t value) noexcept;
     void handleWriteOperations(const SplitWord128& result, uint8_t function, uint8_t offset) noexcept;
 };
 #endif //CHIPSET_TIMERDEVICE_H
